refactor sending emails and added emailTemplates with forgot and reset password emails
